% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allSame.R
\name{allSame}
\alias{allSame}
\title{All Values Same}
\usage{
allSame(x)
}
\arguments{
\item{x}{A vector of values. Can be one of the following objects
classes: \code{numeric}, \code{character}, \code{factor}, or \code{logical}.}
}
\value{
Logical. \code{TRUE} or \code{FALSE}.
}
\description{
Returns a logical as to whether all values \emph{within} a vector
are identical. This function does NOT compare two independent
vectors. Please use \code{\link[=all.equal]{all.equal()}},
for such a purpose combined with \code{\link[=isTRUE]{isTRUE()}}.
}
\examples{
allSame(1:4)
allSame(rep(5, 10))
allSame(rep("A", 10))
allSame(letters)
allSame(c(TRUE, TRUE, TRUE))
allSame(c(TRUE, TRUE, FALSE))
}
\seealso{
\code{\link[=isTRUE]{isTRUE()}}, \code{\link[=all.equal]{all.equal()}}
}
\author{
Stu Field
}
