% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biocDevelActions-package.R
\docType{package}
\name{biocDevelActions-package}
\alias{biocDevelActions}
\alias{biocDevelActions-package}
\title{biocDevelActions: A Test package for BioConductor with Github Actions}
\description{
A test sandbox R package to test Github Actions.
}
\author{
\strong{Maintainer}: Stu Field \email{stu.g.field@gmail.com}

}
\keyword{internal}
\keyword{package}
